feat: clean out the edge case example
