+++ src/test.rs
@@ -59,19 +59,13 @@ where
     }
 }
 
-impl<S, R, T> From<(&[(R, &[S], &[(T, &[(Tau, Tau)])])], &[(S, S)])>
-    for TestTimetable<S, R, T>
+impl<S, R, T> From<(&[(R, &[S], &[(T, &[(Tau, Tau)])])], &[(S, S)])> for TestTimetable<S, R, T>
 where
     S: Ord + Copy,
     R: Ord + Copy + Debug,
     T: Ord + Copy + Debug,
 {
-    fn from(
-        (route_defs, footpath_defs): (
-            &[(R, &[S], &[(T, &[(Tau, Tau)])])],
-            &[(S, S)],
-        ),
-    ) -> Self {
+    fn from((route_defs, footpath_defs): (&[(R, &[S], &[(T, &[(Tau, Tau)])])], &[(S, S)])) -> Self {
         let mut tt = Self::new();
         for &(route_id, stops, trips) in route_defs {
             tt.routes.insert(route_id, stops.to_vec());
@@ -169,25 +163,38 @@ where
     }
 }
 
-/// Recreates the Issue1 network from examples/reconstruction_bugs.rs
-///
-/// Stop mapping: S, A, B, C, D
-/// Route mapping: R1, R2, R3
-/// Trip mapping: R1T1, R2T1, R3Late, R3Early
+/// When a faster route reaches a mid-route stop, the algorithm must record
+/// that stop as the boarding stop — not the earlier stop where the route
+/// scan began. See examples/reboarding.rs for the full network.
 #[test]
-fn reconstruction_bugs_issue1() {
+fn reboarding_picks_correct_boarding_stop() {
     use Route::*;
     use Stop::*;
     use Trip::*;
 
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { S, A, B, C, D }
+    enum Stop {
+        S,
+        A,
+        B,
+        C,
+        D,
+    }
 
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1, R2, R3 }
+    enum Route {
+        R1,
+        R2,
+        R3,
+    }
 
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { R1T1, R2T1, R3Late, R3Early }
+    enum Trip {
+        R1T1,
+        R2T1,
+        R3Late,
+        R3Early,
+    }
 
     let tt = TestTimetable::new()
         .route(R1, &[S, A], &[(R1T1, &[(0, 0), (100, 100)])])
@@ -211,106 +218,135 @@ fn reconstruction_bugs_issue1() {
     assert_eq!(best.plan, vec![(R2, B), (R3, D)]);
 }
 
-/// Same test using the From impl for static declaration
-#[test]
-fn reconstruction_bugs_issue1_from() {
-    use Route::*;
-    use Stop::*;
-    use Trip::*;
-
-    #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { S, A, B, C, D }
-
-    #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1, R2, R3 }
-
-    #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { R1T1, R2T1, R3Late, R3Early }
-
-    let routes: &[(Route, &[Stop], &[(Trip, &[(Tau, Tau)])])] = &[
-        (R1, &[S, A], &[(R1T1, &[(0, 0), (100, 100)])]),
-        (R2, &[S, B], &[(R2T1, &[(0, 0), (30, 30)])]),
-        (
-            R3,
-            &[A, B, C, D],
-            &[
-                (R3Late, &[(105, 105), (110, 110), (120, 120), (130, 130)]),
-                (R3Early, &[(25, 25), (30, 30), (40, 40), (50, 50)]),
-            ],
-        ),
-    ];
-    let tt = TestTimetable::from((routes, &[] as &[_]));
-
-    let journeys = tt.raptor(3, 0, S, D);
-
-    let best = journeys.iter().min_by_key(|j| j.arrival).unwrap();
-    assert_eq!(best.arrival, 50);
-    assert_eq!(best.plan, vec![(R2, B), (R3, D)]);
-}
-
 // ── Edge case tests ─────────────────────────────────────────────────
 
 #[test]
 fn no_journey_disconnected_graph() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B, C, D }
+    enum Stop {
+        A,
+        B,
+        C,
+        D,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1, R2 }
+    enum Route {
+        R1,
+        R2,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1, T2 }
+    enum Trip {
+        T1,
+        T2,
+    }
 
     let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B], &[(Trip::T1, &[(0, 0), (10, 10)])])
-        .route(Route::R2, &[Stop::C, Stop::D], &[(Trip::T2, &[(0, 0), (10, 10)])]);
+        .route(
+            Route::R1,
+            &[Stop::A, Stop::B],
+            &[(Trip::T1, &[(0, 0), (10, 10)])],
+        )
+        .route(
+            Route::R2,
+            &[Stop::C, Stop::D],
+            &[(Trip::T2, &[(0, 0), (10, 10)])],
+        );
 
     let journeys = tt.raptor(3, 0, Stop::A, Stop::D);
-    assert!(journeys.is_empty(), "disconnected graph should yield no journeys");
+    assert!(
+        journeys.is_empty(),
+        "disconnected graph should yield no journeys"
+    );
 }
 
 #[test]
 fn no_journey_missed_connection() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B, C }
+    enum Stop {
+        A,
+        B,
+        C,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1, R2 }
+    enum Route {
+        R1,
+        R2,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1, T2 }
+    enum Trip {
+        T1,
+        T2,
+    }
 
     let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B], &[(Trip::T1, &[(0, 0), (50, 50)])])
-        .route(Route::R2, &[Stop::B, Stop::C], &[(Trip::T2, &[(0, 30), (40, 40)])]);
+        .route(
+            Route::R1,
+            &[Stop::A, Stop::B],
+            &[(Trip::T1, &[(0, 0), (50, 50)])],
+        )
+        .route(
+            Route::R2,
+            &[Stop::B, Stop::C],
+            &[(Trip::T2, &[(0, 30), (40, 40)])],
+        );
 
     let journeys = tt.raptor(3, 0, Stop::A, Stop::C);
-    assert!(journeys.is_empty(), "missed connection should yield no journeys");
+    assert!(
+        journeys.is_empty(),
+        "missed connection should yield no journeys"
+    );
 }
 
 #[test]
 fn no_journey_late_departure() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B }
+    enum Stop {
+        A,
+        B,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1 }
+    enum Route {
+        R1,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1 }
+    enum Trip {
+        T1,
+    }
 
-    let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B], &[(Trip::T1, &[(0, 10), (20, 20)])]);
+    let tt = TestTimetable::new().route(
+        Route::R1,
+        &[Stop::A, Stop::B],
+        &[(Trip::T1, &[(0, 10), (20, 20)])],
+    );
 
     let journeys = tt.raptor(3, 100, Stop::A, Stop::B);
-    assert!(journeys.is_empty(), "late departure should yield no journeys");
+    assert!(
+        journeys.is_empty(),
+        "late departure should yield no journeys"
+    );
 }
 
 #[test]
 fn no_journey_transfers_zero() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B }
+    enum Stop {
+        A,
+        B,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1 }
+    enum Route {
+        R1,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1 }
+    enum Trip {
+        T1,
+    }
 
-    let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B], &[(Trip::T1, &[(0, 0), (10, 10)])]);
+    let tt = TestTimetable::new().route(
+        Route::R1,
+        &[Stop::A, Stop::B],
+        &[(Trip::T1, &[(0, 0), (10, 10)])],
+    );
 
     let journeys = tt.raptor(0, 0, Stop::A, Stop::B);
     assert!(journeys.is_empty(), "transfers=0 should yield no journeys");
@@ -319,32 +355,54 @@ fn no_journey_transfers_zero() {
 #[test]
 fn source_equals_target() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B }
+    enum Stop {
+        A,
+        B,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1 }
+    enum Route {
+        R1,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1 }
+    enum Trip {
+        T1,
+    }
 
-    let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B], &[(Trip::T1, &[(0, 0), (10, 10)])]);
+    let tt = TestTimetable::new().route(
+        Route::R1,
+        &[Stop::A, Stop::B],
+        &[(Trip::T1, &[(0, 0), (10, 10)])],
+    );
 
     let journeys = tt.raptor(3, 0, Stop::A, Stop::A);
-    assert!(journeys.is_empty(), "source == target should yield no journeys");
+    assert!(
+        journeys.is_empty(),
+        "source == target should yield no journeys"
+    );
 }
 
 #[test]
 fn direct_journey_single_route() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B, C }
+    enum Stop {
+        A,
+        B,
+        C,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1 }
+    enum Route {
+        R1,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1 }
+    enum Trip {
+        T1,
+    }
 
-    let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B, Stop::C], &[
-            (Trip::T1, &[(0, 0), (10, 10), (20, 20)]),
-        ]);
+    let tt = TestTimetable::new().route(
+        Route::R1,
+        &[Stop::A, Stop::B, Stop::C],
+        &[(Trip::T1, &[(0, 0), (10, 10), (20, 20)])],
+    );
 
     let journeys = tt.raptor(3, 0, Stop::A, Stop::C);
     assert_eq!(journeys.len(), 1);
@@ -355,15 +413,32 @@ fn direct_journey_single_route() {
 #[test]
 fn direct_journey_picks_fastest_route() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B }
+    enum Stop {
+        A,
+        B,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1, R2 }
+    enum Route {
+        R1,
+        R2,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1, T2 }
+    enum Trip {
+        T1,
+        T2,
+    }
 
     let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B], &[(Trip::T1, &[(0, 0), (100, 100)])])
-        .route(Route::R2, &[Stop::A, Stop::B], &[(Trip::T2, &[(0, 0), (50, 50)])]);
+        .route(
+            Route::R1,
+            &[Stop::A, Stop::B],
+            &[(Trip::T1, &[(0, 0), (100, 100)])],
+        )
+        .route(
+            Route::R2,
+            &[Stop::A, Stop::B],
+            &[(Trip::T2, &[(0, 0), (50, 50)])],
+        );
 
     let journeys = tt.raptor(3, 0, Stop::A, Stop::B);
     let best = journeys.iter().min_by_key(|j| j.arrival).unwrap();
@@ -373,15 +448,33 @@ fn direct_journey_picks_fastest_route() {
 #[test]
 fn exact_time_connection() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B, C }
+    enum Stop {
+        A,
+        B,
+        C,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1, R2 }
+    enum Route {
+        R1,
+        R2,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1, T2 }
+    enum Trip {
+        T1,
+        T2,
+    }
 
     let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B], &[(Trip::T1, &[(0, 0), (20, 20)])])
-        .route(Route::R2, &[Stop::B, Stop::C], &[(Trip::T2, &[(0, 20), (30, 30)])]);
+        .route(
+            Route::R1,
+            &[Stop::A, Stop::B],
+            &[(Trip::T1, &[(0, 0), (20, 20)])],
+        )
+        .route(
+            Route::R2,
+            &[Stop::B, Stop::C],
+            &[(Trip::T2, &[(0, 20), (30, 30)])],
+        );
 
     let journeys = tt.raptor(3, 0, Stop::A, Stop::C);
     assert!(!journeys.is_empty(), "exact-time connection should work");
@@ -393,18 +486,30 @@ fn exact_time_connection() {
 #[test]
 fn multi_trip_picks_earliest_catchable() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B }
+    enum Stop {
+        A,
+        B,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1 }
+    enum Route {
+        R1,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1, T2, T3 }
+    enum Trip {
+        T1,
+        T2,
+        T3,
+    }
 
-    let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B], &[
+    let tt = TestTimetable::new().route(
+        Route::R1,
+        &[Stop::A, Stop::B],
+        &[
             (Trip::T1, &[(0, 5), (15, 15)]),
             (Trip::T2, &[(0, 15), (25, 25)]),
             (Trip::T3, &[(0, 25), (35, 35)]),
-        ]);
+        ],
+    );
 
     // Query at tau=12: T1 departs A@5 (too early), T2 departs A@15 (catchable)
     let journeys = tt.raptor(3, 12, Stop::A, Stop::B);
@@ -415,43 +520,95 @@ fn multi_trip_picks_earliest_catchable() {
 #[test]
 fn two_transfer_journey() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B, C, D }
+    enum Stop {
+        A,
+        B,
+        C,
+        D,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1, R2, R3 }
+    enum Route {
+        R1,
+        R2,
+        R3,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1, T2, T3 }
+    enum Trip {
+        T1,
+        T2,
+        T3,
+    }
 
     let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B], &[(Trip::T1, &[(0, 0), (10, 10)])])
-        .route(Route::R2, &[Stop::B, Stop::C], &[(Trip::T2, &[(0, 10), (20, 20)])])
-        .route(Route::R3, &[Stop::C, Stop::D], &[(Trip::T3, &[(0, 20), (30, 30)])]);
+        .route(
+            Route::R1,
+            &[Stop::A, Stop::B],
+            &[(Trip::T1, &[(0, 0), (10, 10)])],
+        )
+        .route(
+            Route::R2,
+            &[Stop::B, Stop::C],
+            &[(Trip::T2, &[(0, 10), (20, 20)])],
+        )
+        .route(
+            Route::R3,
+            &[Stop::C, Stop::D],
+            &[(Trip::T3, &[(0, 20), (30, 30)])],
+        );
 
     let journeys = tt.raptor(3, 0, Stop::A, Stop::D);
     assert!(!journeys.is_empty());
     let best = journeys.iter().min_by_key(|j| j.arrival).unwrap();
     assert_eq!(best.arrival, 30);
-    assert_eq!(best.plan, vec![
-        (Route::R1, Stop::B),
-        (Route::R2, Stop::C),
-        (Route::R3, Stop::D),
-    ]);
+    assert_eq!(
+        best.plan,
+        vec![
+            (Route::R1, Stop::B),
+            (Route::R2, Stop::C),
+            (Route::R3, Stop::D),
+        ]
+    );
 }
 
 #[test]
 fn pareto_optimal_fewer_transfers_vs_faster() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B, D }
+    enum Stop {
+        A,
+        B,
+        D,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1, R2, R3 }
+    enum Route {
+        R1,
+        R2,
+        R3,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1, T2, T3 }
+    enum Trip {
+        T1,
+        T2,
+        T3,
+    }
 
     let tt = TestTimetable::new()
         // Direct slow route A→D
-        .route(Route::R1, &[Stop::A, Stop::D], &[(Trip::T1, &[(0, 0), (200, 200)])])
+        .route(
+            Route::R1,
+            &[Stop::A, Stop::D],
+            &[(Trip::T1, &[(0, 0), (200, 200)])],
+        )
         // Fast 2-leg: A→B via R2, B→D via R3
-        .route(Route::R2, &[Stop::A, Stop::B], &[(Trip::T2, &[(0, 0), (40, 40)])])
-        .route(Route::R3, &[Stop::B, Stop::D], &[(Trip::T3, &[(0, 40), (100, 100)])]);
+        .route(
+            Route::R2,
+            &[Stop::A, Stop::B],
+            &[(Trip::T2, &[(0, 0), (40, 40)])],
+        )
+        .route(
+            Route::R3,
+            &[Stop::B, Stop::D],
+            &[(Trip::T3, &[(0, 40), (100, 100)])],
+        );
 
     let journeys = tt.raptor(3, 0, Stop::A, Stop::D);
     assert_eq!(journeys.len(), 2, "should have 2 pareto-optimal journeys");
@@ -469,15 +626,34 @@ fn pareto_optimal_fewer_transfers_vs_faster() {
 #[test]
 fn footpath_enables_connection() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B, C, D }
+    enum Stop {
+        A,
+        B,
+        C,
+        D,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1, R2 }
+    enum Route {
+        R1,
+        R2,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1, T2 }
+    enum Trip {
+        T1,
+        T2,
+    }
 
     let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B], &[(Trip::T1, &[(0, 0), (10, 10)])])
-        .route(Route::R2, &[Stop::C, Stop::D], &[(Trip::T2, &[(0, 20), (30, 30)])])
+        .route(
+            Route::R1,
+            &[Stop::A, Stop::B],
+            &[(Trip::T1, &[(0, 0), (10, 10)])],
+        )
+        .route(
+            Route::R2,
+            &[Stop::C, Stop::D],
+            &[(Trip::T2, &[(0, 20), (30, 30)])],
+        )
         .footpath(Stop::B, Stop::C)
         .transfer_time(Stop::B, Stop::C, 5);
 
@@ -485,39 +661,74 @@ fn footpath_enables_connection() {
     // NOTE: The algorithm correctly propagates arrival times through footpaths,
     // but journey reconstruction can't trace back through footpath-only transfers
     // (no boarding tree entry for the footpath destination). This documents that limitation.
-    assert!(journeys.is_empty(), "footpath-only transfer not reconstructable");
+    assert!(
+        journeys.is_empty(),
+        "footpath-only transfer not reconstructable"
+    );
 }
 
 #[test]
 fn footpath_transfer_time_causes_miss() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B, C, D }
+    enum Stop {
+        A,
+        B,
+        C,
+        D,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1, R2 }
+    enum Route {
+        R1,
+        R2,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1, T2 }
+    enum Trip {
+        T1,
+        T2,
+    }
 
     let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B], &[(Trip::T1, &[(0, 0), (50, 50)])])
-        .route(Route::R2, &[Stop::C, Stop::D], &[(Trip::T2, &[(0, 52), (60, 60)])])
+        .route(
+            Route::R1,
+            &[Stop::A, Stop::B],
+            &[(Trip::T1, &[(0, 0), (50, 50)])],
+        )
+        .route(
+            Route::R2,
+            &[Stop::C, Stop::D],
+            &[(Trip::T2, &[(0, 52), (60, 60)])],
+        )
         .footpath(Stop::B, Stop::C)
         .transfer_time(Stop::B, Stop::C, 5); // 50+5=55 > 52
 
     let journeys = tt.raptor(3, 0, Stop::A, Stop::D);
-    assert!(journeys.is_empty(), "footpath transfer time should cause miss");
+    assert!(
+        journeys.is_empty(),
+        "footpath transfer time should cause miss"
+    );
 }
 
 #[test]
 fn early_termination_no_improvement() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B }
+    enum Stop {
+        A,
+        B,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1 }
+    enum Route {
+        R1,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1 }
+    enum Trip {
+        T1,
+    }
 
-    let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B], &[(Trip::T1, &[(0, 0), (10, 10)])]);
+    let tt = TestTimetable::new().route(
+        Route::R1,
+        &[Stop::A, Stop::B],
+        &[(Trip::T1, &[(0, 0), (10, 10)])],
+    );
 
     let j1 = tt.raptor(1, 0, Stop::A, Stop::B);
     let j100 = tt.raptor(100, 0, Stop::A, Stop::B);
@@ -532,15 +743,32 @@ fn early_termination_no_improvement() {
 #[test]
 fn dominance_prunes_slower_arrival() {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Stop { A, B }
+    enum Stop {
+        A,
+        B,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Route { R1, R2 }
+    enum Route {
+        R1,
+        R2,
+    }
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum Trip { T1, T2 }
+    enum Trip {
+        T1,
+        T2,
+    }
 
     let tt = TestTimetable::new()
-        .route(Route::R1, &[Stop::A, Stop::B], &[(Trip::T1, &[(0, 0), (50, 50)])])
-        .route(Route::R2, &[Stop::A, Stop::B], &[(Trip::T2, &[(0, 0), (100, 100)])]);
+        .route(
+            Route::R1,
+            &[Stop::A, Stop::B],
+            &[(Trip::T1, &[(0, 0), (50, 50)])],
+        )
+        .route(
+            Route::R2,
+            &[Stop::A, Stop::B],
+            &[(Trip::T2, &[(0, 0), (100, 100)])],
+        );
 
     let journeys = tt.raptor(3, 0, Stop::A, Stop::B);
     // Both routes are discovered in round 1, so the slower one is dominated
